refactor: move LoadSurveyResult dependencies to a protocols file
