+++ src/data/usecases/survey-result/load-survey-result/db-load-survey-result-protocols.ts
@@ -1,0 +1,3 @@
+export * from '@/domain/usecases/survey-result/load-survey-result'
+export * from '@/domain/models/survey-result'
+export * from '@/data/protocols/db/survey-result/load-survey-result-repository'
