+++ src/data/usecases/survey-result/load-survey-result/db-load-survey-result.spec.ts
@@ -1,6 +1,6 @@
 import { DbLoadSurveyResult } from './db-load-survey-result'
+import { LoadSurveyResultRepository } from './db-load-survey-result-protocols'
 import { mockLoadSurveyResultRepository } from '@/data/test'
-import { LoadSurveyResultRepository } from '@/data/protocols/db/survey-result/load-survey-result-repository'
 
 type SutTypes = {
   sut: DbLoadSurveyResult
