+++ src/data/usecases/survey-result/load-survey-result/db-load-survey-result.ts
@@ -1,6 +1,4 @@
-import { LoadSurveyResultRepository } from '@/data/protocols/db/survey-result/load-survey-result-repository'
-import { LoadSurveyResult } from '@/domain/usecases/survey-result/load-survey-result'
-import { SurveyResultModel } from '@/domain/models/survey-result'
+import { LoadSurveyResultRepository, SurveyResultModel, LoadSurveyResult } from './db-load-survey-result-protocols'
 
 export class DbLoadSurveyResult implements LoadSurveyResult {
   constructor (private readonly loadSurveyResultRepository: LoadSurveyResultRepository) {}
